requirements.txt

google-auth>=2.43.0
google-auth-httplib2>=0.2.1
google-auth-oauthlib>=1.2.2
gspread>=5.12.4
pandas>=2.3.3
plotly>=6.4.0
st-gsheets-connection>=0.1.0
streamlit>=1.51.0
streamlit-extras>=0.6.0
mysql-connector-python
